requirements.txt
azure-quantum
numpy
scipy
pandas
matplotlib
pycryptodome
tqdm
rich
colorama
scapy 